requirements.txt
numpy
python-dotenv
pandas
pymongo
certifi
pymongo[srv]
dill
pyaml
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart

-e .